Test the root and search controllers
I've also moved those controllers to a module in the controller
subdirectory for coherence.

In the search controller, the code guards against search results where
the identifiers would be None (uid for users and cn for groups) but I
have not found a way to reproduce this situation for testing. I'm not
sure if this use case can really happen, but I've left the protection
there in case it actually happens.
